% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{gla_compute_solar_positions}
\alias{gla_compute_solar_positions}
\title{Compute solar positions and radiation values}
\usage{
gla_compute_solar_positions(
  lat_deg,
  long_deg,
  elev,
  clearsky_coef,
  time_step_min,
  day_start,
  day_end,
  day_res,
  elev_res,
  azi_res,
  solar_constant
)
}
\description{
Compute solar positions and radiation values
}
\keyword{internal}
