% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{gla_load_points}
\alias{gla_load_points}
\title{Load points}
\usage{
gla_load_points(x, dem, ...)
}
\arguments{
\item{x}{Either a file path to any file that \code{sf::read_sf} can read, or an
sf object containing point geometries}

\item{dem}{Either a file path to a DEM raster file, or a \code{SpatRaster} object}

\item{...}{Additional arguments passed to \code{sf::read_sf()} when \code{x} is a file path}
}
\description{
Load points
}
