% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{gla_process_fisheye_photo_single}
\alias{gla_process_fisheye_photo_single}
\title{Process single fisheye photo and calculate SSR}
\usage{
gla_process_fisheye_photo_single(
  solar_data,
  img_file,
  lat_deg,
  long_deg,
  Kt,
  elev_res,
  azi_res,
  rotation_deg,
  keep_gap_fraction_data = FALSE,
  radial_distortion = "equidistant",
  threshold = 0
)
}
\description{
Process single fisheye photo and calculate SSR
}
\keyword{internal}
