% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar.R
\name{gla_transform_lidar}
\alias{gla_transform_lidar}
\title{Process lidar data for hemispherical photo creation}
\usage{
gla_transform_lidar(
  las_input,
  x_meters,
  y_meters,
  elev_m,
  cam_ht,
  min_dist = 1
)
}
\description{
Process lidar data for hemispherical photo creation
}
\keyword{internal}
