% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaplightr-package.R
\docType{package}
\name{gaplightr-package}
\alias{gaplightr}
\alias{gaplightr-package}
\title{gaplightr: Gap Light Analysis of Virtual and Real Fisheye Photography}
\description{
Analyzes forest canopy gap light transmission using LiDAR point cloud data and hemispherical photography. Provides tools for processing LiDAR data, computing horizon angles, and calculating gap light metrics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hakaiinstitute.github.io/gaplightr/}
  \item \url{https://github.com/HakaiInstitute/gaplightr}
  \item Report bugs at \url{https://github.com/HakaiInstitute/gaplightr/issues}
}

}
\author{
\strong{Maintainer}: Sam Albers \email{sam.albers@hakai.org}

Authors:
\itemize{
  \item Gord Frazer \email{gordonwfrazer@gmail.com} (Conceptualization and original code)
}

Other contributors:
\itemize{
  \item Hakai Institute [copyright holder]
}

}
\keyword{internal}
