% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{gla_load_points}
\alias{gla_load_points}
\title{Load points}
\usage{
gla_load_points(path, dem_path)
}
\arguments{
\item{path}{path to a any files that \code{sf::read_sf} can read}

\item{dem_path}{a dem path}
}
\description{
Load points
}
